Liked 관련 코드 typescript로 리팩토링
