+++ src/hooks/useNotification.ts
@@ -5,7 +5,6 @@ import { alarmState } from '../recoil/apiDataAtoms';
 const useNotification = (alarmMsg: string) => {
   const [noti, setNoti] = useRecoilState(alarmState);
 
-  // console.log('알람메시지 잘 전달되었나?', alarmMsg);
   const deleteNoti = () => {
     setNoti((prev) => prev.slice(1));
   };
@@ -17,11 +16,6 @@ const useNotification = (alarmMsg: string) => {
     setTimeout(() => {
       setNoti((prev) => prev.slice(1));
     }, 2000);
-    // useEffect(()=>{
-    //   return () => {
-    //     clearTimeout(timer)
-    //   }
-    // },[setNoti])
   };
   return { noti, setNoti, addNoti, deleteNoti };
 };
