+++ src/types/apiDataTypes.ts
@@ -1,0 +1,51 @@
+export interface DetailDataTypes {
+  contentid: string;
+  contenttypeid: string;
+  title: string;
+  createdtime: string;
+  modifiedtime: string;
+  tel: string;
+  telname: string;
+  homepage: string;
+  booktour: string;
+  firstimage: string;
+  firstimage2: string;
+  areacode: string;
+  sigungucode: string;
+  cat1: string;
+  cat2: string;
+  cat3: string;
+  addr1: string;
+  addr2: string;
+  zipcode: string;
+  mapx: string;
+  mapy: string;
+  mlevel: string;
+  overview: string;
+}
+
+// export interface spotDetailDataTypes {
+//   contentid: string;
+//   contenttypeid: string;
+//   title: string;
+//   createdtime: string;
+//   modifiedtime: string;
+//   tel: string;
+//   telname: string;
+//   homepage: string;
+//   booktour: string;
+//   firstimage: string;
+//   firstimage2: string;
+//   areacode: string;
+//   sigungucode: string;
+//   cat1: string;
+//   cat2: string;
+//   cat3: string;
+//   addr1: string;
+//   addr2: string;
+//   zipcode: string;
+//   mapx: string;
+//   mapy: string;
+//   mlevel: string;
+//   overview: string;
+// }
